test HELP

+++ src/main/java/org/example/functions/AbstractTabulatedFunction.java
@@ -1,7 +1,7 @@
 package org.example.functions;
 
 abstract class AbstractTabulatedFunction  {
-    protected abstract double floorIndexOfX(double x);
+    protected abstract int floorIndexOfX(double x);
     protected abstract double extrapolateLeft(double x);
     protected abstract double extrapolateRight(double x);
     protected abstract double interpolate(double x, int floorIndex);
